Added with-screen-switch hoc and Switch
