+++ package.json
@@ -37,6 +37,7 @@
     "react": "16.8.6",
     "react-dom": "16.8.6",
     "react-redux": "7.0.3",
+    "react-router-dom": "5.0.1",
     "recompose": "0.30.0",
     "redux": "4.0.1",
     "redux-thunk": "2.3.0",
